+++ docs/release_notes.rst
@@ -14,6 +14,7 @@ New Features
 * Add support for `property_decorators` config to ignore D401.
 * Add support for Python 3.10 (#554).
 * Replace D10X errors with D419 if docstring exists but is empty (#559).
+* Add `ignore-functions` option to allow some functions to skip docstring checking (#587).
 
 Bug Fixes